tests: refactor db cleanup
remove from url, to avoid disasters
