+++ site/BusinessLogic/BusinessLogic.csproj
@@ -108,7 +108,6 @@
     </Compile>
     <Compile Include="Bases\GenericMoveRepository.cs" />
     <Compile Include="Current.cs" />
-    <Compile Include="Truncate\Cleaner.cs" />
     <Compile Include="Helpers\MoveExtension.cs" />
     <Compile Include="Helpers\ThemeExtension.cs" />
     <Compile Include="InterfacesAndBases\IPasswordForm.cs" />
@@ -171,7 +170,6 @@
     <Compile Include="Services\RobotService.cs" />
     <Compile Include="Services\SafeService.cs" />
     <Compile Include="Response\SignInInfo.cs" />
-    <Compile Include="Truncate\Entities.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Authentication\Authentication.csproj">
