+++ site/Tests/Core/BusinessLogic/MainStep.cs
@@ -4,9 +4,9 @@ using DFM.BusinessLogic;
 using DFM.BusinessLogic.Exceptions;
 using DFM.BusinessLogic.Repositories.Mappings;
 using DFM.BusinessLogic.Response;
-using DFM.BusinessLogic.Truncate;
 using DFM.Entities;
 using DFM.Language;
+using DFM.Tests.Helpers;
 using Keon.NHibernate.Base;
 using NUnit.Framework;
 using TechTalk.SpecFlow;
